Use docker ps -n 1 to check docker health
Under heavy load, `docker ps` takes quite a while to return. That means that you may end up restarting docker daemon and killing a bunch of pods making the load heavier.
Using `docker ps -n 1` allow to check that the daemon is running, it is more lightweight and it should take less time.